Enhancement: Add english-words-alpha.txt collection
Book source: https://github.com/touchtypie/english-words

Performance will be very bad if the book is going to be 400k lines long. This should not be part of book practice, but instead part of another kind of practice e.g. word practice, which is not consistent with the [philosophy of touchtypie](https://github.com/touchtypie/touchtypie#philosophy).